pass down svg props + move to icons folder
