+++ packages/runner-ct/src/app/app.tsx
@@ -12,10 +12,10 @@ import Iframes from '../iframe/iframes'
 import Message from '../message/message'
 import './app.scss'
 import { ReporterHeader } from './ReporterHeader'
-import { BurgerIcon } from './BurgerIcon'
 import EventManager from '../lib/event-manager'
 import { Hidden } from '../lib/Hidden'
 import { SpecList } from '../SpecList'
+import { BurgerIcon } from '../icons/Burger'
 import { ResizableBox } from '../lib/ResizableBox'
 import { useWindowSize } from '../lib/useWindowSize'
