+++ packages/runner-ct/src/icons/ExpandGroup.tsx
@@ -1,12 +1,13 @@
 import React from 'react'
 
-export const ExpandGroup: React.FC = () => {
+export const ExpandGroup: React.FC<React.SVGProps<SVGSVGElement>> = (props) => {
   return (
     <svg
       width="11"
       height="11"
-      viewBox="0 0 11 11"
+      {...props}
       fill="none"
+      viewBox="0 0 11 11"
       xmlns="http://www.w3.org/2000/svg"
     >
       <path
